chore: rename select dir
